Move wasmtime's instantiation test into a separate file.
